Refactor : ResponseEntity 적용

+++ src/main/java/com/challnege/delivery/domain/review/service/ReviewService.java
@@ -5,11 +5,15 @@ import com.challnege.delivery.domain.member.repository.MemberRepository;
 import com.challnege.delivery.domain.restaurant.entity.Restaurant;
 import com.challnege.delivery.domain.restaurant.repository.RestaurantRepository;
 import com.challnege.delivery.domain.review.dto.ReviewRequestDto;
+import com.challnege.delivery.domain.review.dto.ReviewResponseDto;
 import com.challnege.delivery.domain.review.entity.Review;
 import com.challnege.delivery.domain.review.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -20,22 +24,35 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
 
     @Transactional
-    public void postReview(Long restaurantId, ReviewRequestDto requestDto) {
+    public ResponseEntity<ReviewResponseDto> postReview(Long restaurantId, ReviewRequestDto requestDto) {
         Restaurant restaurant = restaurantRepository.findById(restaurantId).orElseThrow(
                 () -> new IllegalArgumentException("해당 업장은 존재하지 않습니다.")
         );
         Member member = memberRepository.findById(requestDto.getMember().getMemberId()).orElseThrow(
-                () -> new IllegalArgumentException()
+                () -> new IllegalArgumentException("해당 회원은 존재하지 않습니다.")
         );
-        reviewRepository.save(new Review(restaurant,member,requestDto));
+
+        Review review = new Review(restaurant, member, requestDto);
+        reviewRepository.save(review);
+
+        ReviewResponseDto responseDto = new ReviewResponseDto(review);
+        return ResponseEntity.ok(responseDto);
     }
 
     @Transactional
-    public void deleteReview(Long restaurantsId, Long reviewId) {
-        Restaurant restaurant = restaurantRepository.findById(restaurantsId).orElseThrow(
+    public ResponseEntity<ReviewResponseDto> deleteReview(Long restaurantId, Long reviewId) {
+        Restaurant restaurant = restaurantRepository.findById(restaurantId).orElseThrow(
                 () -> new IllegalArgumentException("해당 업장은 존재하지 않습니다.")
         );
-        Review review = reviewRepository.findByRestaurantAndReviewId(restaurant,reviewId);
+
+        Review review = reviewRepository.findByRestaurantAndReviewId(restaurant, reviewId);
+        if (review == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 리뷰는 존재하지 않습니다.");
+        }
+        ReviewResponseDto responseDto = new ReviewResponseDto(review);
+
         reviewRepository.delete(review);
+
+        return ResponseEntity.ok(responseDto);
     }
 }
